feat(attendance): enhance attendance handling and add tests
Streamlined attendance-related route definitions and improved type handling for request inputs. Refactored policies to simplify logic and added missing authorization checks for approval actions. Introduced extensive unit tests for services and repositories to ensure correctness and better coverage.

+++ app/Modules/Attendance/Models/Attendance.php
@@ -168,6 +168,6 @@ class Attendance extends Model
         $totalMinutes = $this->check_in->diffInMinutes($this->check_out);
         $breakMinutes = $this->breaks()->sum('duration_minutes') ?? 0;
 
-        return $totalMinutes - $breakMinutes;
+        return (int) ($totalMinutes - $breakMinutes);
     }
 }
